move components into /addon dir, various fixes
