+++ addon/components/each-property/component.js
@@ -1,0 +1,27 @@
+import Ember from 'ember';
+
+export function getPropertyInputType(property) {
+  if (property.validValues && Array.isArray(property.validValues)) {
+    return 'select';
+  }
+
+  if (property.type === 'boolean') {
+    return 'radio';
+  }
+
+  return 'text';
+}
+
+export default Ember.Component.extend({
+  propertyCollection: Ember.computed('properties.[]', function() {
+    let propertyHash = this.get('properties');
+    let propertyKeys = Object.keys(propertyHash);
+
+    return propertyKeys.map((key) => {
+      let property = propertyHash[key];
+
+      return { key, property, type: getPropertyInputType(property),
+               childProperties: property.properties };
+    });
+  })
+});
