+++ app/components/each-property/component.js
@@ -1,34 +1,1 @@
-import Ember from 'ember';
-
-export function getPropertyInputType(property) {
-  if(property.validValues && Array.isArray(property.validValues)) {
-    return 'select';
-  }
-
-  switch(property.type) {
-    case 'boolean':
-      return 'radio';
-    case 'string':
-      return 'text';
-    case 'number':
-      return 'number';
-    case 'integer':
-      return 'number';
-  }
-
-  return 'text';
-}
-
-export default Ember.Component.extend({
-  propertyCollection: Ember.computed('properties.[]', function() {
-    let propertyHash = this.get('properties');
-    let propertyKeys = Object.keys(propertyHash);
-
-    return propertyKeys.map((key) => {
-      let property = propertyHash[key];
-
-      return { key, property, type: getPropertyInputType(property),
-               childProperties: property.properties }
-    });
-  })
-});
+export { default } from 'ember-json-schema/components/each-property/component';
